% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bootStrap}
\alias{bootStrap}
\title{Bootstrap Resampling}
\usage{
bootStrap(data, column, n)
}
\arguments{
\item{data}{Dataframe where you want to take the bootstrap from}

\item{column}{Column of interest. This must be numeric}

\item{n}{Number of bootstrap resamples you want to take.}
}
\value{
Returns a 95\% confidence interval (quartile) as well as a histogram of the spread of the bootstraps
}
\description{
\code{bootStrap} performs simple bootstrapping on a numeric dataset.
}
\examples{
bootStrap(beaver1, 2, 200)
bootStrap(beaver2, 2, 10000)

}
