% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev.R
\docType{data}
\name{fev}
\alias{fev}
\title{FEV data}
\format{An object of class \code{data.frame} with 654 rows and 7 columns.}
\usage{
data(fev)
}
\description{
This is FEV data from our 514 class
}
\keyword{datasets}
