% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Again.R
\name{MyMath}
\alias{MyMath}
\title{To perform a simple math equation.}
\usage{
MyMath(x)
}
\arguments{
\item{x}{Input vector, must be a character vector.}
}
\value{
6*x+7
}
\description{
MyMath multiplies a number by 6 and adds 7.
}
